refactor: :recycle: use new notation of Bun 1.1.5
