+++ src/templates/notes/script.js
