+++ src/templates/notes/styles.css
