+++ src/templates/present/index.ts
@@ -1,6 +1,6 @@
-import layoutHTML from "./layout.html.txt";
-import scriptJS from "./script.js.txt";
-import stylesCSS from "./styles.css.txt";
+import layoutHTML from "./layout.html" with { type: "text" };
+import scriptJS from "./script.js" with { type: "text" };
+import stylesCSS from "./styles.css" with { type: "text" };
 
 export const view = layoutHTML;
 export const script = scriptJS;
